refactor: consolidate key managers into sdkClient for single-file setup
Tutorial users now only need to save one file (sdkClient.mjs) instead of
three to get started with write operations. IdentityKeyManager.mjs and
AddressKeyManager.mjs become thin re-exports for backward compatibility.
